chore: extract realtime from app.py
