+++ projects/fal/src/fal/_typing.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from typing import Any, Callable, TypeVar
+
+EndpointT = TypeVar("EndpointT", bound=Callable[..., Any])
